ESD-1269: Build-tag glamour out of the WASM build
